Import global store instead of useState to refresh page
The issue here was that when we get it from the store directly we dont have any language set at the start of the app, because the state is zero. In this way we have the initial local state language in the app. Theres still an issue that the selected language does not appear at selected if the user logs out (should be fixable with getting the state)